Repository/code/twitter-scraper/requirements.txt
elasticsearch
langdetect
ruamel.yaml
tweepy